+++ CHANGELOG.md
@@ -15,9 +15,11 @@
 - #1139 Adding centralized task executor for kernels
 - #1188 Implement upper/lower operators
 - #1193 Implement string REPLACE
+- #1218 Added smiles test set
 - #1201 Implement string TRIM
 - #1216 Add unit test for DAYOFWEEK
 - #1205 Implement string REVERSE
+- #1220 Implement string LEFT and RIGHT 
 
 
 ## Improvements